Refactor tests to global folder
